Extract some code from Controllers::Filesystem
