+++ src/lib/y2partitioner/actions/controllers.rb
@@ -34,3 +34,4 @@ require "y2partitioner/actions/controllers/filesystem"
 require "y2partitioner/actions/controllers/md"
 require "y2partitioner/actions/controllers/add_partition"
 require "y2partitioner/actions/controllers/lvm_lv"
+require "y2partitioner/actions/controllers/encryption"
