+++ src/lib/y2partitioner/actions/controllers/filesystem.rb
@@ -36,7 +36,6 @@ Yast.import "Stage"
 #   * Remove a filesystem
 #   * Edit format options
 #   * Edit partition id
-#   * Change encryption
 #
 # but it also includes logic related to a filesystem:
 #   * Set mount point
@@ -63,9 +62,6 @@ module Y2Partitioner
 
         # @return [Boolean] Whether the user wants to encrypt the device
         attr_accessor :encrypt
-
-        # @return [String] Password for the encryption device
-        attr_accessor :encrypt_password
 
         # Name of the plain device
         #
@@ -145,15 +141,6 @@ module Y2Partitioner
           return false if filesystem.nil?
 
           new?(filesystem)
-        end
-
-        # Whether a new encryption device will be created for the block device
-        #
-        # @return [Boolean]
-        def to_be_encrypted?
-          return false unless can_change_encrypt?
-
-          encrypt && !blk_device.encrypted?
         end
 
         # Mount point of the current filesystem
@@ -347,34 +334,6 @@ module Y2Partitioner
           after_change_mount_point
         end
 
-        # Applies last changes to the block device at the end of the wizard, which
-        # mainly means
-        #
-        #   * removing unused LvmPv descendant (bsc#1129663)
-        #   * encrypting the device or removing the encryption layer for non preexisting devices.
-        def finish
-          return unless can_change_encrypt?
-
-          remove_unused_lvm_pv
-
-          if to_be_encrypted?
-            blk_device.encrypt(password: encrypt_password)
-          elsif blk_device.encrypted? && !encrypt
-            blk_device.remove_encryption
-          end
-        ensure
-          @restorer.update_checkpoint
-        end
-
-        # Removes from the block device or its encryption layer a LvmPv not associated to an LvmVg
-        # (bsc#1129663)
-        def remove_unused_lvm_pv
-          device = blk_device.encryption || blk_device
-          lvm_pv = device.lvm_pv
-
-          device.remove_descendants if lvm_pv&.descendants&.none?
-        end
-
         # Whether is possible to define the generic format options for the current
         # filesystem
         #
@@ -470,6 +429,13 @@ module Y2Partitioner
           filesystem.supports_btrfs_subvolumes?
         end
 
+        # Saves the current status of the block device
+        #
+        # @see BlkDeviceRestorer#update_checkpoint
+        def update_checkpoint
+          @restorer.update_checkpoint
+        end
+
         private
 
         def working_graph
@@ -478,10 +444,6 @@ module Y2Partitioner
 
         def system_graph
           DeviceGraphs.instance.system
-        end
-
-        def can_change_encrypt?
-          filesystem.nil? || new?(filesystem)
         end
 
         def new?(device)
