+++ test/y2partitioner/actions/controllers/filesystem_test.rb
@@ -124,55 +124,6 @@ describe Y2Partitioner::Actions::Controllers::Filesystem do
     end
   end
 
-  describe "#to_be_encrypted?" do
-    context "when the currently editing device has a filesystem that existed previously" do
-      it "returns false" do
-        expect(subject.to_be_encrypted?).to eq(false)
-      end
-    end
-
-    context "when the currently editing device does not have a filesystem that existed previously" do
-      before do
-        allow(subject).to receive(:encrypt).and_return(encrypt)
-        allow(device).to receive(:encrypted?).and_return(encrypted)
-        allow(device).to receive(:filesystem).and_return(filesystem)
-        allow(subject).to receive(:blk_device).and_return(device)
-      end
-
-      let(:encrypt) { false }
-      let(:encrypted) { false }
-      let(:filesystem) { nil }
-
-      context "and the device has not been marked to encrypt" do
-        let(:encrypt) { false }
-
-        it "returns false" do
-          expect(subject.to_be_encrypted?).to eq(false)
-        end
-      end
-
-      context "and the device has been marked to encrypt" do
-        let(:encrypt) { true }
-
-        context "and the device is currently encrypted" do
-          let(:encrypted) { true }
-
-          it "returns false" do
-            expect(subject.to_be_encrypted?).to eq(false)
-          end
-        end
-
-        context "and the device is not currently encrypted" do
-          let(:encrypted) { false }
-
-          it "returns true" do
-            expect(subject.to_be_encrypted?).to eq(true)
-          end
-        end
-      end
-    end
-  end
-
   describe "#mount_point" do
     context "when the currently editing device has a filesystem" do
       context "and the filesystem has a mount point" do
@@ -1220,594 +1171,6 @@ describe Y2Partitioner::Actions::Controllers::Filesystem do
             include_examples "btrfs subvolumes check"
           end
         end
-      end
-    end
-  end
-
-  describe "#finish" do
-    before do
-      allow(subject).to receive(:can_change_encrypt?).and_return(can_change_encrypt)
-    end
-
-    context "when it is not possible to change the encrypt" do
-      let(:can_change_encrypt) { false }
-
-      it "does nothing" do
-        devicegraph = Y2Partitioner::DeviceGraphs.instance.current.dup
-        subject.finish
-
-        expect(devicegraph).to eq(Y2Partitioner::DeviceGraphs.instance.current)
-      end
-    end
-
-    context "when it is possible to change the encrypt" do
-      let(:can_change_encrypt) { true }
-
-      before do
-        allow(subject).to receive(:encrypt).and_return(encrypt)
-        allow(subject).to receive(:encrypt_password).and_return(password)
-      end
-
-      let(:encrypt) { false }
-      let(:password) { "12345678" }
-
-      context "and the device was already encrypted" do
-        before do
-          device.remove_descendants
-          encryption = device.create_encryption("foo")
-          encryption.create_filesystem(Y2Storage::Filesystems::Type::EXT4)
-        end
-
-        context "and it is marked to be encrypted" do
-          let(:encrypt) { true }
-
-          it "does nothing" do
-            devicegraph = Y2Partitioner::DeviceGraphs.instance.current.dup
-            subject.finish
-
-            expect(devicegraph).to eq(Y2Partitioner::DeviceGraphs.instance.current)
-          end
-        end
-
-        context "and it is not marked to be encrypted" do
-          let(:encrypt) { false }
-
-          it "removes the encryption" do
-            expect(subject.blk_device.encryption).to_not be_nil
-            subject.finish
-            expect(subject.blk_device.encryption).to be_nil
-          end
-        end
-      end
-
-      context "and the device is not encrypted" do
-        context "and it is marked to be encrypted" do
-          let(:encrypt) { true }
-
-          it "encrypts the device" do
-            subject.finish
-            expect(subject.blk_device.encryption).to_not be_nil
-            expect(subject.blk_device.encryption.password).to eq(password)
-          end
-        end
-
-        context "and it is not marked to be encrypted" do
-          let(:encrypt) { false }
-
-          it "does nothing" do
-            devicegraph = Y2Partitioner::DeviceGraphs.instance.current.dup
-            subject.finish
-
-            expect(devicegraph).to eq(Y2Partitioner::DeviceGraphs.instance.current)
-          end
-        end
-      end
-    end
-
-    context "when the device has an unused LvmPv" do
-      let(:scenario) { "unused_lvm_pvs.xml" }
-
-      let(:can_change_encrypt) { true }
-      let(:encrypt) { false }
-      let(:password) { "12345678" }
-
-      before do
-        allow(subject).to receive(:encrypt).and_return(encrypt)
-        allow(subject).to receive(:encrypt_password).and_return(password)
-      end
-
-      context "which was not encrypted" do
-        let(:dev_name) { "/dev/sda2" }
-
-        context "and it is marked to be encrypted" do
-          let(:encrypt) { true }
-
-          it "removes the unused LvmPv" do
-            subject.finish
-            expect(subject.blk_device.lvm_pv).to be_nil
-          end
-
-          it "encrypts the device" do
-            subject.finish
-            expect(subject.blk_device.encryption).to_not be_nil
-            expect(subject.blk_device.encryption.password).to eq(password)
-          end
-        end
-
-        context "and it is not marked to be encrypted" do
-          let(:encrypt) { false }
-
-          it "removes the unused LvmPv" do
-            subject.finish
-            expect(subject.blk_device.lvm_pv).to be_nil
-          end
-        end
-      end
-
-      context "which was already encrypted" do
-        let(:dev_name) { "/dev/sda3" }
-
-        context "and it is marked to keep the encryption" do
-          let(:encrypt) { true }
-
-          it "keeps the encryption" do
-            encryption = subject.blk_device.encryption
-            subject.finish
-            expect(subject.blk_device.encryption).to eq(encryption)
-          end
-
-          it "removes the unused LvmPv" do
-            subject.finish
-            expect(subject.blk_device.lvm_pv).to be_nil
-          end
-        end
-
-        context "and it is not marked to be encrypted" do
-          let(:encrypt) { false }
-
-          it "removes the encryption" do
-            subject.finish
-            expect(subject.blk_device.encryption).to be_nil
-          end
-
-          it "removes the unused LvmPv" do
-            subject.finish
-            expect(subject.blk_device.lvm_pv).to be_nil
-          end
-        end
-      end
-    end
-  end
-
-  describe "#format_options_supported?" do
-    context "when the currently editing device has no filesystem" do
-      let(:dev_name) { "/dev/sdb7" }
-
-      it "returns false" do
-        expect(controller.format_options_supported?).to eq false
-      end
-    end
-
-    context "when the currently editing device has a filesystem" do
-      context "that is a preexisting Btrfs one" do
-        let(:dev_name) { "/dev/sda2" }
-
-        it "returns false" do
-          expect(controller.format_options_supported?).to eq false
-        end
-      end
-
-      context "that is a preexisting no-Btrfs one" do
-        let(:dev_name) { "/dev/sdb5" }
-
-        it "returns false" do
-          expect(controller.format_options_supported?).to eq false
-        end
-      end
-
-      context "that is a new (to be created) Btrfs" do
-        it "returns false" do
-          device.remove_descendants
-          device.create_filesystem(Y2Storage::Filesystems::Type::BTRFS)
-
-          expect(controller.format_options_supported?).to eq false
-        end
-      end
-
-      context "that is a new (to be created) no-Btrfs" do
-        it "returns true" do
-          device.remove_descendants
-          device.create_filesystem(Y2Storage::Filesystems::Type::EXT4)
-
-          expect(controller.format_options_supported?).to eq true
-        end
-      end
-    end
-  end
-
-  describe "#snapshots_supported?" do
-    before { allow(Yast::Mode).to receive(:installation).and_return inst_mode }
-    let(:inst_mode) { false }
-
-    context "when the currently editing device has no filesystem" do
-      let(:dev_name) { "/dev/sdb7" }
-
-      context "in installation mode" do
-        let(:inst_mode) { true }
-
-        it "returns false" do
-          expect(controller.snapshots_supported?).to eq false
-        end
-      end
-
-      context "in normal mode" do
-        let(:inst_mode) { false }
-
-        it "returns false" do
-          expect(controller.snapshots_supported?).to eq false
-        end
-      end
-    end
-
-    context "when the currently editing device has a filesystem" do
-      context "that is a preexisting Btrfs one" do
-        let(:dev_name) { "/dev/sda2" }
-
-        context "in installation mode" do
-          let(:inst_mode) { true }
-
-          it "returns false" do
-            expect(controller.snapshots_supported?).to eq false
-          end
-        end
-
-        context "in normal mode" do
-          let(:inst_mode) { false }
-
-          it "returns false" do
-            expect(controller.snapshots_supported?).to eq false
-          end
-        end
-      end
-
-      context "that is a preexisting no-Btrfs one" do
-        let(:dev_name) { "/dev/sdb5" }
-
-        context "in installation mode" do
-          let(:inst_mode) { true }
-
-          it "returns false" do
-            expect(controller.snapshots_supported?).to eq false
-          end
-        end
-
-        context "in normal mode" do
-          let(:inst_mode) { false }
-
-          it "returns false" do
-            expect(controller.snapshots_supported?).to eq false
-          end
-        end
-      end
-
-      context "that is a new (to be created) Btrfs" do
-        before do
-          device.remove_descendants
-          device.create_filesystem(Y2Storage::Filesystems::Type::BTRFS)
-          device.filesystem.mount_path = mntpnt
-        end
-
-        context "if is going to be mounted as root" do
-          let(:mntpnt) { "/" }
-
-          context "in installation mode" do
-            let(:inst_mode) { true }
-
-            it "returns true" do
-              expect(controller.snapshots_supported?).to eq true
-            end
-          end
-
-          context "in normal mode" do
-            let(:inst_mode) { false }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-        end
-
-        context "if is not going to be mounted as root" do
-          let(:mntpnt) { "/var" }
-
-          context "in installation mode" do
-            let(:inst_mode) { true }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-
-          context "in normal mode" do
-            let(:inst_mode) { false }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-        end
-      end
-
-      context "that is a new (to be created) no-Btrfs" do
-        before do
-          device.remove_descendants
-          device.create_filesystem(Y2Storage::Filesystems::Type::EXT4)
-          device.filesystem.mount_path = mntpnt
-        end
-
-        context "if is going to be mounted as root" do
-          let(:mntpnt) { "/" }
-
-          context "in installation mode" do
-            let(:inst_mode) { true }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-
-          context "in normal mode" do
-            let(:inst_mode) { false }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-        end
-
-        context "if is not going to be mounted as root" do
-          let(:mntpnt) { "/var" }
-
-          context "in installation mode" do
-            let(:inst_mode) { true }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-
-          context "in normal mode" do
-            let(:inst_mode) { false }
-
-            it "returns false" do
-              expect(controller.snapshots_supported?).to eq false
-            end
-          end
-        end
-      end
-    end
-  end
-
-  describe "#mount_paths" do
-    before do
-      allow(storage_arch).to receive(:efiboot?).and_return efi
-      allow(Yast::Stage).to receive(:initial).and_return installation
-    end
-
-    let(:storage_arch) { instance_double(Storage::Arch) }
-
-    let(:system_mountpoints) { ["/", "/var", "/opt"] }
-    let(:additional_mountpoints) { ["/home", "/srv", "/tmp", "/usr/local"] }
-
-    context "if no mount points are assigned so far" do
-      let(:scenario) { "windows-linux-multiboot-pc" }
-      before do
-        devicegraph.filesystems.each do |fs|
-          fs.remove_mount_point unless fs.mount_point.nil?
-        end
-      end
-
-      RSpec.shared_examples "always included" do
-        it "includes the additional mount points" do
-          expect(controller.mount_paths).to include(*additional_mountpoints)
-        end
-
-        context "if the filesystem is of type SWAP" do
-          let(:dev_name) { "/dev/sda2" }
-
-          it "includes 'swap' as the first element of the list" do
-            expect(controller.mount_paths.first).to eq "swap"
-          end
-        end
-
-        context "if the filesystem is not of type SWAP" do
-          let(:dev_name) { "/dev/sda1" }
-
-          it "does not include 'swap'" do
-            expect(controller.mount_paths).to_not include("swap")
-          end
-        end
-      end
-
-      context "during installation" do
-        let(:installation) { true }
-
-        context "in a s390 system" do
-          let(:architecture) { :s390 }
-          let(:efi) { false }
-
-          it "includes the system mount points" do
-            expect(controller.mount_paths).to include(*system_mountpoints)
-          end
-
-          it "includes /boot and /boot/zipl" do
-            expect(controller.mount_paths).to include("/boot", "/boot/zipl")
-          end
-
-          it "does not include /boot/efi" do
-            expect(controller.mount_paths).to_not include("/boot/efi")
-          end
-
-          include_examples "always included"
-        end
-
-        context "in a non-s390 system" do
-          let(:architecture) { :x86 }
-
-          context "with EFI boot" do
-            let(:efi) { true }
-
-            it "includes the system mount points" do
-              expect(controller.mount_paths).to include(*system_mountpoints)
-            end
-
-            it "includes /boot and /boot/efi" do
-              expect(controller.mount_paths).to include("/boot", "/boot/efi")
-            end
-
-            it "does not include /boot/zipl" do
-              expect(controller.mount_paths).to_not include("/boot/zipl")
-            end
-
-            include_examples "always included"
-          end
-
-          context "with no EFI" do
-            let(:efi) { false }
-
-            it "includes the system mount points" do
-              expect(controller.mount_paths).to include(*system_mountpoints)
-            end
-
-            it "includes /boot" do
-              expect(controller.mount_paths).to include("/boot")
-            end
-
-            it "does not include /boot/zipl or /boot/efi" do
-              expect(controller.mount_paths).to_not include("/boot/efi")
-              expect(controller.mount_paths).to_not include("/boot/zipl")
-            end
-
-            include_examples "always included"
-          end
-        end
-      end
-
-      context "in an installed system" do
-        let(:installation) { false }
-
-        context "if it's a s390 system" do
-          let(:architecture) { :s390 }
-
-          let(:efi) { false }
-
-          it "does not include the system mount points" do
-            system_mountpoints.each do |mp|
-              expect(controller.mount_paths).to_not include mp
-            end
-          end
-
-          it "does not include /boot, /boot/efi or /boot/zipl" do
-            expect(controller.mount_paths).to_not include("/boot")
-            expect(controller.mount_paths).to_not include("/boot/zipl")
-            expect(controller.mount_paths).to_not include("/boot/efi")
-          end
-
-          include_examples "always included"
-        end
-
-        context "if it's not a s390 system" do
-          let(:architecture) { :x86 }
-
-          context "with EFI boot" do
-            let(:efi) { true }
-
-            it "does not include the system mount points" do
-              system_mountpoints.each do |mp|
-                expect(controller.mount_paths).to_not include mp
-              end
-            end
-
-            it "does not include /boot, /boot/efi or /boot/zipl" do
-              expect(controller.mount_paths).to_not include("/boot")
-              expect(controller.mount_paths).to_not include("/boot/zipl")
-              expect(controller.mount_paths).to_not include("/boot/efi")
-            end
-
-            include_examples "always included"
-          end
-
-          context "with no EFI" do
-            let(:efi) { false }
-
-            it "does not include the system mount points" do
-              system_mountpoints.each do |mp|
-                expect(controller.mount_paths).to_not include mp
-              end
-            end
-
-            it "does not include /boot, /boot/efi or /boot/zipl" do
-              expect(controller.mount_paths).to_not include("/boot")
-              expect(controller.mount_paths).to_not include("/boot/zipl")
-              expect(controller.mount_paths).to_not include("/boot/efi")
-            end
-
-            include_examples "always included"
-          end
-        end
-      end
-    end
-
-    context "if some mount points are already taken" do
-      let(:scenario) { "mixed_disks" }
-      # Let's enforce the temptative presence of /
-      let(:installation) { true }
-      # Let's enforce the temptative presence of swap
-      let(:dev_name) { "/dev/sdb1" }
-      # These two are not much relevant; swap, /home and / should be there in
-      # any combination
-      let(:architecture) { :x86 }
-
-      let(:efi) { false }
-
-      it "does not include the already mounted regular paths" do
-        expect(controller.mount_paths).to_not include("/")
-        expect(controller.mount_paths).to_not include("/home")
-      end
-
-      it "includes 'swap' despite it being already used" do
-        expect(controller.mount_paths).to include("swap")
-      end
-    end
-  end
-
-  describe "#suggested_mount_path" do
-    context "with an unmounted ESP" do
-      let(:scenario) { "efi_not_mounted" }
-      let(:dev_name) { "/dev/sda1" }
-
-      it "suggests /boot/efi as the mount point" do
-        expect(subject.suggested_mount_path).to be == "/boot/efi"
-        expect(subject.mount_paths.first).to be == "/boot/efi"
-      end
-    end
-
-    context "with swap partitions" do
-      let(:scenario) { "swaps" }
-      let(:dev_name) { "/dev/sda2" }
-
-      it "suggests swap as the mount point" do
-        expect(subject.suggested_mount_path).to be == "swap"
-        expect(subject.mount_paths.first).to be == "swap"
-      end
-    end
-
-    context "for other partitions" do
-      let(:scenario) { "mixed_disks" }
-      let(:dev_name) { "/dev/sda2" }
-
-      it "does not suggests a mount point" do
-        expect(subject.suggested_mount_path).to be_nil
       end
     end
   end
